requirements.txt
websockets>=12.0
watchdog>=4.0
google-api-python-client>=2.100
google-auth>=2.25
google-auth-oauthlib>=1.2
paramiko>=3.4
cryptography>=42.0
GitPython>=3.1
pyyaml>=6.0
bcrypt>=4.1
